+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Travis-CI badge entry in `Cargo.toml`
+- Warning in example
+
 
 ## [0.1.1] - 2017-09-15
 ### Added