+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2017-09-17
 ### Added
 - Added method overview in documentation
 - `StableVec::contains()`
@@ -30,7 +32,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Travis-CI badge entry in `Cargo.toml`
 - Warning in example
 
-
 ## [0.1.1] - 2017-09-15
 ### Added
 - Added this `CHANGELOG.md`
@@ -49,6 +50,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Everything.
 
 
-[Unreleased]: https://github.com/LukasKalbertodt/stable-vec/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/LukasKalbertodt/stable-vec/compare/v0.2.0...HEAD
+[0.2.ß]: https://github.com/LukasKalbertodt/stable-vec/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/LukasKalbertodt/stable-vec/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/LukasKalbertodt/stable-vec/compare/v0.1.0...v0.1.1